// Code/src/WebUI/ClientApp/src/app/SiteManager.V4-api.ts
export class AppRoleDto implements IAppRoleDto {
    appRoleId?: number;
    name?: string | undefined;

    constructor(data?: IAppRoleDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.appRoleId = _data["appRoleId"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): AppRoleDto {
        data = typeof data === 'object' ? data : {};
        let result = new AppRoleDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["appRoleId"] = this.appRoleId;
        data["name"] = this.name;
        return data; 
    }
}

export interface IAppRoleDto {
    appRoleId?: number;
    name?: string | undefined;
}

export class CreateRoleCommand implements ICreateRoleCommand {
    roleName?: string | undefined;

    constructor(data?: ICreateRoleCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.roleName = _data["roleName"];
        }
    }

    static fromJS(data: any): CreateRoleCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateRoleCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["roleName"] = this.roleName;
        return data; 
    }
}

export interface ICreateRoleCommand {
    roleName?: string | undefined;
}
// ...
export class TodosVm implements ITodosVm {
    priorityLevels?: PriorityLevelDto[] | undefined;
    lists?: TodoListDto[] | undefined;

    constructor(data?: ITodosVm) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["priorityLevels"])) {
                this.priorityLevels = [] as any;
                for (let item of _data["priorityLevels"])
                    this.priorityLevels!.push(PriorityLevelDto.fromJS(item));
            }
            if (Array.isArray(_data["lists"])) {
                this.lists = [] as any;
                for (let item of _data["lists"])
                    this.lists!.push(TodoListDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): TodosVm {
        data = typeof data === 'object' ? data : {};
        let result = new TodosVm();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.priorityLevels)) {
            data["priorityLevels"] = [];
            for (let item of this.priorityLevels)
                data["priorityLevels"].push(item.toJSON());
        }
        if (Array.isArray(this.lists)) {
            data["lists"] = [];
            for (let item of this.lists)
                data["lists"].push(item.toJSON());
        }
        return data; 
    }
}
// ...
export class TodoListDto implements ITodoListDto {
    id?: number;
    title?: string | undefined;
    items?: TodoItemDto[] | undefined;

    constructor(data?: ITodoListDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.title = _data["title"];
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(TodoItemDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): TodoListDto {
        data = typeof data === 'object' ? data : {};
        let result = new TodoListDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item.toJSON());
        }
        return data; 
    }
}

export interface ITodoListDto {
    id?: number;
    title?: string | undefined;
    items?: TodoItemDto[] | undefined;
}

export class TodoItemDto implements ITodoItemDto {
    id?: number;
    listId?: number;
    title?: string | undefined;
    done?: boolean;
    priority?: number;
    note?: string | undefined;

    constructor(data?: ITodoItemDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.listId = _data["listId"];
            this.title = _data["title"];
            this.done = _data["done"];
            this.priority = _data["priority"];
            this.note = _data["note"];
        }
    }

    static fromJS(data: any): TodoItemDto {
        data = typeof data === 'object' ? data : {};
        let result = new TodoItemDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["listId"] = this.listId;
        data["title"] = this.title;
        data["done"] = this.done;
        data["priority"] = this.priority;
        data["note"] = this.note;
        return data; 
    }
}
// ...
export class AppUserDto implements IAppUserDto {
    appUserId?: number;
    userName?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    emailConfirmed?: boolean;
    appRoles?: string[] | undefined;

    constructor(data?: IAppUserDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.appUserId = _data["appUserId"];
            this.userName = _data["userName"];
            this.email = _data["email"];
            this.phoneNumber = _data["phoneNumber"];
            this.emailConfirmed = _data["emailConfirmed"];
            if (Array.isArray(_data["appRoles"])) {
                this.appRoles = [] as any;
                for (let item of _data["appRoles"])
                    this.appRoles!.push(item);
            }
        }
    }

    static fromJS(data: any): AppUserDto {
        data = typeof data === 'object' ? data : {};
        let result = new AppUserDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["appUserId"] = this.appUserId;
        data["userName"] = this.userName;
        data["email"] = this.email;
        data["phoneNumber"] = this.phoneNumber;
        data["emailConfirmed"] = this.emailConfirmed;
        if (Array.isArray(this.appRoles)) {
            data["appRoles"] = [];
            for (let item of this.appRoles)
                data["appRoles"].push(item);
        }
        return data; 
    }
}

export interface IAppUserDto {
    appUserId?: number;
    userName?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    emailConfirmed?: boolean;
    appRoles?: string[] | undefined;
}

export class CreateUserCommand implements ICreateUserCommand {
    userName?: string | undefined;
    password?: string | undefined;
    confirmPassword?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    appRoles?: string[] | undefined;
    sendRegistrationConfirmation?: boolean;
    confirmationUrl?: string | undefined;

    constructor(data?: ICreateUserCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userName = _data["userName"];
            this.password = _data["password"];
            this.confirmPassword = _data["confirmPassword"];
            this.email = _data["email"];
            this.phoneNumber = _data["phoneNumber"];
            if (Array.isArray(_data["appRoles"])) {
                this.appRoles = [] as any;
                for (let item of _data["appRoles"])
                    this.appRoles!.push(item);
            }
            this.sendRegistrationConfirmation = _data["sendRegistrationConfirmation"];
            this.confirmationUrl = _data["confirmationUrl"];
        }
    }

    static fromJS(data: any): CreateUserCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateUserCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userName"] = this.userName;
        data["password"] = this.password;
        data["confirmPassword"] = this.confirmPassword;
        data["email"] = this.email;
        data["phoneNumber"] = this.phoneNumber;
        if (Array.isArray(this.appRoles)) {
            data["appRoles"] = [];
            for (let item of this.appRoles)
                data["appRoles"].push(item);
        }
        data["sendRegistrationConfirmation"] = this.sendRegistrationConfirmation;
        data["confirmationUrl"] = this.confirmationUrl;
        return data; 
    }
}

export interface ICreateUserCommand {
    userName?: string | undefined;
    password?: string | undefined;
    confirmPassword?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    appRoles?: string[] | undefined;
    sendRegistrationConfirmation?: boolean;
    confirmationUrl?: string | undefined;
}

export class UpdateUserCommand implements IUpdateUserCommand {
    appUserId?: number;
    userName?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    resetPassword?: boolean;
    appRoles?: string[] | undefined;
    passwordResetUrl?: string | undefined;

    constructor(data?: IUpdateUserCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.appUserId = _data["appUserId"];
            this.userName = _data["userName"];
            this.email = _data["email"];
            this.phoneNumber = _data["phoneNumber"];
            this.resetPassword = _data["resetPassword"];
            if (Array.isArray(_data["appRoles"])) {
                this.appRoles = [] as any;
                for (let item of _data["appRoles"])
                    this.appRoles!.push(item);
            }
            this.passwordResetUrl = _data["passwordResetUrl"];
        }
    }

    static fromJS(data: any): UpdateUserCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateUserCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["appUserId"] = this.appUserId;
        data["userName"] = this.userName;
        data["email"] = this.email;
        data["phoneNumber"] = this.phoneNumber;
        data["resetPassword"] = this.resetPassword;
        if (Array.isArray(this.appRoles)) {
            data["appRoles"] = [];
            for (let item of this.appRoles)
                data["appRoles"].push(item);
        }
        data["passwordResetUrl"] = this.passwordResetUrl;
        return data; 
    }
}